% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corn}
\alias{corn}
\title{CDL corn classes}
\format{An object of class \code{numeric} of length 6.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
corn
}
\description{
An array of CDL enumerations that contain corn.  The corn enumeration contains: 
\itemize{
\item 1 - Corn
\item 225 - Double Crop, Winter Wheat and Corn
\item 226 - Double Crop, Oats and Corn
\item 237 - Double Crop, Barley and Corn 
\item 241 - Double Crop, Corn and Soybeans
\item 251 - Non-Irrigated Corn
}
}
\keyword{datasets}

