% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{CDL cotton classes}
\format{An object of class \code{numeric} of length 4.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
cotton
}
\description{
An array of CDL enumerations that contain cotton.  The cotton enumeration contains: 
\itemize{
\item 2 - Cotton
\item 232 - Double Crop, Lettuce and Cotton
\item 238 - Double Crop, Winter Wheat and Cotton
\item 239 - Double Crop, Soybeans and Cotton
}
}
\keyword{datasets}
