% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countyFIPS_t.R
\name{countyFIPS}
\alias{countyFIPS}
\title{Get US county FIPS code}
\usage{
countyFIPS(x)
}
\arguments{
\item{x}{A character string. county and state names seperated by  ",". For the county, it should be full name. For the state, it can be the two-letter postal abbreviation or the full name. The string is case insensitive.}
}
\value{
5-digit county FIPS code
}
\description{
Retrieve 5-digit US county FIPS code by names
}
\examples{
countyFIPS("fairfax county,va")
}

