% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createComparableCDL.R
\name{createComparableCDL}
\alias{createComparableCDL}
\title{Create comparable raster images.}
\usage{
createComparableCDL(rasterList, filenames, baseIndex, progress = "")
}
\arguments{
\item{A}{vector of raster images.}

\item{A}{vector of file names of raster images to coerce into a raster list, 
if rasterList is not provided.}

\item{is}{the index of the vector element that all other elements will be converted
into.}

\item{is}{a string for the raster progress bar type, default "" is none, 
"text" provides text output, "window" provides a gui window if available.}
}
\value{
A list of raster images matching in extent, resolution and projection.
}
\description{
\code{createComparableCDL} uses a base index within a raster list, and sets all other
raster images within the list to the same resolution, projection, and extent.  The 
raster function resample is used to change projections and extents, therefore this 
function may be quite slow without tuning.
}
\examples{
r <- getCDL('rhode island',2008:2010)
r2 <- createComparableCDL(r,baseIndex=1)
}
\author{
Jonathan Lisic, \email{jlisic@gmail.com}
}
