% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{durumWheat}
\alias{durumWheat}
\title{CDL durum wheat classes}
\format{An object of class \code{numeric} of length 3.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
durumWheat
}
\description{
An array of CDL enumerations that contain durum wheat.  The durum wheat enumeration contains: 
\itemize{
\item 22 - Durum Wheat 
\item 230 - Double Crop Lettuce and Durum Wheat
\item 234 - Double Crop Durum Wheat and Sorghum 
}
}
\keyword{datasets}
