% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips.R
\name{fips}
\alias{fips}
\title{FIPS code conversion function.}
\usage{
fips(x, to = "FIPS")
}
\arguments{
\item{x}{A character string or numeric FIPS code. Character input can be the two-letter 
postal abbreviation, the full name of a state, or a FIPS code in character format. 
The string is case insensitive.  FIPS codes are the only numeric input supported.}

\item{to}{A character string of output type:  "FIPS" will return a numeric fips code.
"Abbreviation" will return a two letter state abbreviation.  "Name" will return the
full state name with spaces.  The default output is a numeric FIPS code.}
}
\value{
The output type specified by the "to" argument.  If no match can be made, the 
 program returns NA.
}
\description{
\code{fips} converts U.S. state names and abbreviations to and from FIPS codes.
}
\details{
The Federal Information Processing Standard (FIPS) provides a set of standard numeric 
codes for refering to U.S. states.  This function converts between FIPS codes, state two 
letter abbreviations, and full state names.
}
\examples{
fips("ia")
fips('northcarolina', to='Abbreviation')
fips('North Carolina')
fips(44,to='Name')
}

