% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCDL_bbox.R
\name{getCDL_bbox}
\alias{getCDL_bbox}
\title{Get CDL raster data}
\usage{
getCDL_bbox(x, year)
}
\arguments{
\item{x}{A numeric matrix. Bounding box coordinates of an interested couty. first column has the minimum values, the second has the maximum values; rows represent the spatial dimensions.}

\item{year}{A numerical vector. A set of interestedy years}
}
\value{
a list of CDL raster objects of interested county for a set of years.
}
\description{
Retrieve CDL data for a set of years by bounding box
}
\examples{
getCDL_fips("01001",c(2013,2014))
}

