% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getCDL_fips.R
\name{getCDL_fips}
\alias{getCDL_fips}
\title{Get CDL raster data}
\usage{
getCDL_fips(x, year, FIPS = TRUE)
}
\arguments{
\item{x}{A character string. either 5-digit county FIPS code or 2-digit state FIPS code if type= or county and state names seperated by ",".}

\item{year}{A numerical vector. A set of interestedy years}

\item{FIPS}{logical object. If TRUE, x is 5-digit county FIPS code. Otherwise, x is a character string to match in the county and state names. Default is TRUE.}
}
\value{
a list of CDL raster objects of interested county for a set of years.
}
\description{
Retrieve CDL data for a set of years by names or 5-digit county FIPS code
}
\examples{
getCDL_fips("01001",c(2013,2014))
getCDL_fips("autauga county,al",c(2013,2014),FIPS = FALSE)
}

