% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNames.R
\name{getNames}
\alias{getNames}
\title{Retrieve the names from state and county fips code.}
\usage{
getNames(x, y)
}
\arguments{
\item{x}{A character string. Two-digit state fips code.}

\item{y}{A character string. Three-digit state fips code.}
}
\value{
A list of corresponding state name and county name.
}
\description{
Retrieve the names from state and county fips code.
}
\examples{
getNames("19","015")
}

