% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nothing}
\alias{nothing}
\title{CDL nothing class}
\format{An object of class \code{numeric} of length 1.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
nothing
}
\description{
An array of CDL enumerations that contain the nothing class.  The nothing enumeration contains: 
\itemize{
\item 0 - Background 
}
}
\keyword{datasets}

