% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pasture}
\alias{pasture}
\title{CDL pasture classes}
\format{
An object of class \code{numeric} of length 5.
}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
pasture
}
\description{
An array of CDL enumerations that contain pasture.  The pasture enumeration contains: 
\itemize{
\item 37 - Other Hay/Non Alfalfa 
\item 38 - Camelina 
\item 39 - Buckwheat 
\item 62 - Pasture/Grass 
\item 171 - Grassland Herbaceous 
}
}
\keyword{datasets}
