% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{projCDL}
\alias{projCDL}
\title{The default projection of CDL data}
\format{An object of class \code{character} of length 1.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
projCDL
}
\description{
The proj4 string used for all CDL data.
 "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83 
  +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
}
\keyword{datasets}

