% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soybeans}
\alias{soybeans}
\title{CDL soybeans classes}
\format{An object of class \code{numeric} of length 9.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
soybeans
}
\description{
An array of CDL enumerations that contain soybeans.  The soybeans enumeration contains: 
\itemize{
\item 5 - Soybeans 
\item 26 - Double Crop Winter Wheat and Soybeans 
\item 239 - Double Crop Soybeans and Cotton 
\item 240 - Double Crop Soybeans and Oats 
\item 241 - Double Crop Corn and Soybeans 
\item 252 - Non-Irrigated Soybeans 
\item 254 - Double Crop Barley and Soybeans 
\item 254 - Double Crop Barley and Soybeans 
\item 255 - Non-Irrigated Double Crop Winter Wheat and Soybeans 
}
}
\keyword{datasets}
