% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{springWheat}
\alias{springWheat}
\title{CDL spring wheat classes}
\format{An object of class \code{numeric} of length 1.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
springWheat
}
\description{
An array of CDL enumerations that contain spring wheat.  The spring wheat enumeration contains: 
\itemize{
\item 23 - Spring Wheat 
}
}
\keyword{datasets}
