% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stateFIPS.R
\name{stateFIPS}
\alias{stateFIPS}
\title{Get US state FIPS code}
\usage{
stateFIPS(x)
}
\arguments{
\item{x}{A character string. State names. It can be the two-letter postal abbreviation or the full name. The string is case insensitive.}
}
\value{
2-digit state FIPS code
}
\description{
Retrieve 2-digit US state FIPS code by names
}
\examples{
stateFIPS("ia")
}

