% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stateNames.R
\docType{data}
\name{stateNames}
\alias{stateNames}
\title{US State Names and FIPS codes}
\format{A data frame with 54 observations on the following 3 variables:

\code{STATE} a character vector. State Postal Code

\code{STATENAME} a character vector.State Name

\code{STATEFP} a character vector.State FIPS Code}
\source{
\url{https://www.census.gov/geo/reference/ansi_statetables.html}
}
\usage{
data(stateNames)
}
\description{
State Postal Code,Names and FIPS code
}
\examples{
data(stateNames)
}
\keyword{datasets}

