% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateNamesCDL.R
\name{updateNamesCDL}
\alias{updateNamesCDL}
\title{Label CDL classes.}
\usage{
updateNamesCDL(y)
}
\arguments{
\item{y}{A numeric array of integers associated with CDL categories.}
}
\value{
An array of strings labeling each CDL class.  If the CDL class is 
 unspecified then the original integer is returned.
}
\description{
\code{updateNamesCDL} converts numeric CDL categories to class labels.
}
\examples{
updateNamesCDL(0:255)
}

