% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/UScensusFIPS.R
\docType{data}
\name{UScensusFIPS}
\alias{UScensusFIPS}
\title{US county FIPS codes}
\format{A data frame with 3235 observations on the following 6 variables:

\code{STATE} a character vector. State Postal Code

\code{STATEFP} a character vector.State FIPS Code

\code{STATENAME} a character vector.State Name

\code{COUNTYFP} a character vector.County FIPS Code

\code{COUNTYNAME} a character vector.County Name and Legal/Statistical Area Description

\code{CLASSFP} a character vector.FIPS Class Code}
\source{
\url{https://www.census.gov/geo/reference/codes/cou.html}
}
\usage{
data(UScensusFIPS)
}
\description{
Data from 2010 US census, including state postal code, state FIPS code, county FIPS code, county name,state full name and FIPS class code.
}
\examples{
data(UScensusFIPS)
}
\keyword{datasets}

