% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{varNamesCDL}
\alias{varNamesCDL}
\title{Enumerated CDL classes}
\format{An object of class \code{character} of length 278.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
varNamesCDL
}
\description{
A list of enumerated CDL classes and class descriptions.
}
\keyword{datasets}
