% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{CDL water classes}
\format{An object of class \code{numeric} of length 2.}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
water
}
\description{
An array of CDL enumerations that contain water.  The water enumeration contains: 
\itemize{
\item 83 - Water 
\item 111 - Open Water 
}
}
\keyword{datasets}
