% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{winterWheat}
\alias{winterWheat}
\title{CDL winter wheat classes}
\format{
An object of class \code{numeric} of length 7.
}
\source{
\url{https://www.nass.usda.gov/Research_and_Science/Cropland/metadata/meta.php}
}
\usage{
winterWheat
}
\description{
An array of CDL enumerations that contain winter wheat.  The winter wheat enumeration contains: 
\itemize{
\item 24 - Winter Wheat 
\item 26 - Double Crop Winter Wheat and Soybeans 
\item 225 - Double Crop Winter Wheat and Corn 
\item 236 - Double Crop Winter Wheat and Sorghum 
\item 238 - Double Crop Winter Wheat and Cotton 
\item 253 - Non-Irrigated Winter Wheat 
\item 255 - Non-Irrigated Double Crop Winter Wheat and Soybeans 
}
}
\keyword{datasets}
