\name{cdlTools-package}
\alias{cdlTools-package}
\alias{cdlTools}
\docType{package}
\title{
Cropland Data Layer Tools
}
\description{
Tools to obtain and use CDL data with the raster package.  
}
\details{
\tabular{ll}{
Package: \tab cdlTools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-26\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Lu Chen and Jonathan Lisic 
Maintainer: Jonathan Lisic <jonathan.lisic@nass.usda.gov>
}
\references{
Boryan, Claire, et al. "Monitoring US agriculture:  the US department of agriculture, national agricultural statistics service, cropland data layer program." Geocarto International 26.5 (2011): 341-358.
}
\keyword{ package }
\keyword{ datasets }
\seealso{
  raster
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
