% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stateNames}
\alias{stateNames}
\title{U.S. Census 2010 State FIPS Data}
\format{An object of class \code{data.frame} with 54 rows and 3 columns.}
\source{
\url{http://www2.census.gov/geo/docs/reference/codes/files/national_county.txt}
}
\usage{
stateNames
}
\description{
U.S. Census 2010 State FIPS Data containing names, FIPS codes, and abbreviations.
}
\details{
\describe{
\item{STATE}{State two letter abbreviation}
\item{STATENAME}{State name}
\item{STATEFP}{State FIPS code}
}
}
\keyword{datasets}

